requirements.txt
coinbase-advancedtrade-python==0.5.0
flask==2.3.3
flask-wtf==1.1.1
apscheduler==3.10.1
python-dotenv==1.0.0
gunicorn==20.1.0
